% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCDLData.R
\name{GetCDLData}
\alias{GetCDLData}
\title{Access CDL raster data}
\usage{
GetCDLData(aoi = NULL, year = NULL, type = NULL, mat = FALSE, crs = NULL)
}
\arguments{
\item{aoi}{Area of interest. Could be a county FIPS code, a coordinate with two numeric values, multiple coordinates (e.g, three coordinates that defines a triangle),
or four values that defines a rectangle. The default coordinate system (used by CDL) is the Albers equal-area conic projection, or Albers projection. Users can provide
coordinates from a different projection method, but user have to specify the coordinate system in the \code{crs} argument.
For example, users can provide longitude/latitude coordinates here, while letting \code{crs} be '+init=epsg:4326'.}

\item{year}{year of data to request. Can be a numerical value or a character.}

\item{type}{type of aoi. 'f' for county, 'ps' for points, 'b' for box, 'p' for a single point.}

\item{mat}{TRUE/FALSE. If FALSE (default), return a raster file; if TRUE, return a matrix that saves the data.}

\item{crs}{projection system for the coordinate, such as '+init=epsg:4326' for longitude/latitude.}
}
\value{
The function returns a raster file or a matrix that saves the land use information. The file that links the category codes with the category names
are available at~\url{https://www.nass.usda.gov/Research_and_Science/Cropland/docs/cdl_codes_names.xlsx}.
}
\description{
A function that retrieves CDL raster data for any area of interests for a year.
}
\examples{

# Example 1. Retrieve data for the Champaign county in Illinois (FIPS = 17109) in 2018.
data <- GetCDLData(aoi = 17019, year = 2018, type = 'f')
raster::plot(data) # plot the data.

# Example 2. Retrieve data for a single point by long/lat in 2018.
data <- GetCDLData(aoi = c(-94.6754,42.1197), year = 2018, type = 'p', crs = '+init=epsg:4326')
data
# Below uses the same point, but under the default coordinate system
data <- GetCDLData(aoi = c(108777,2125055), year = 2018, type = 'p')
data

# Example 3. Retrieve data for a triangle defined by three coordinates in 2018.
data <- GetCDLData(aoi = c(175207,2219600,175207,2235525,213693,2219600), year = 2018, type = 'ps')
raster::plot(data)

# Example 4. Retrieve data for a rectangle box defined by three corner points in 2018.
data <- GetCDLData(aoi = c(130783,2203171,153923,2217961), year = '2018', type = 'b')
raster::plot(data)

}
