% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRot.R
\name{plotRot}
\alias{plotRot}
\title{Generate land use transition data for area of interest.}
\usage{
plotRot(data, top = 5)
}
\arguments{
\item{data}{Land use change data generated by the \code{rotate} function.}

\item{top}{Top land use types.}
}
\value{
The function returns a ggplot object.
}
\description{
A function that generate land use transitions from year1 to year2 for area of interest.
}
